sepolicy_vndr: Rename exported3_radio_prop to radio_control_prop
The context name exported3_radio_prop is ambiguous and does not reflect the usage and role of the properties. This changes its name to
radio_control_prop.

Some downstream branches are still using exported3_radio_prop, so get_prop(domain, radio_control_prop) is added to avoid regression. It's
just a workaround and to be removed soon, after all exported3_radio_prop are cleaned up.

Bug: 162214733
Test: boot a device with a sim and see basic functions work
Change-Id: I6937ffbc26cfef5a117bf9351f70d35dc5f44b17